requirements.txt
Flask==0.10.1
Frozen-Flask==0.11
Jinja2==2.7.2
Markdown==2.4
MarkupSafe==0.19
PyYAML==3.10
Werkzeug==0.9.4
boto==2.28.0
ipython==1.2.1
itsdangerous==0.23
pyenchant==1.6.5
wsgiref==0.1.2
